add concepts from example app

+++ servers/ruby/entry.rb
@@ -16,14 +16,22 @@ module Front::App; end
 module Front::Components; end
 module Front::Services; end
 
+DONT_LOAD_REGEXP = /migration/
+
 service_names = Dir['../../{app,components,services}/**/*.rb'].map{ |file|
-  require_relative file
-  "front/#{file[6...-3]}" # rm ../../ and .rb and add front/
-}
+  unless file =~ DONT_LOAD_REGEXP
+    require_relative file
+    "front/#{file[6...-3]}" # rm ../../ and .rb and add front/
+  end
+}.compact
 
 services = service_names.reduce({}){ |h, name|
   service = name.camelize.constantize.new
   h[name] = service; h
+}
+
+Dir['../../{app,components,services}/**/migration'].map{ |migration|
+  require_relative migration
 }
 
 NAME = 0
